+++ frontend/src/stores/userStore.ts
@@ -265,7 +265,7 @@ export const userStore = defineStore("user", function () {
     return { firstTime: isFirstTime, isTwoFAEnabled: user.isTwoFAEnabled };
   }
 
-  async function updateProfile() {
+  async function updateProfile() : Promise<boolean> {
     let body = {} as any;
     body.avatar = user.avatar;
     body.image = user.image;
@@ -277,24 +277,20 @@ export const userStore = defineStore("user", function () {
       headers: { Authorization: `Bearer ${user.access_token_server}` },
       body: new URLSearchParams(body),
     };
-    await fetch(env.BACKEND_SERVER_URL + "/users/update_profile", options)
+    const response = await fetch(env.BACKEND_SERVER_URL + "/users/update_profile", options)
       .then(async function (response: any) {
         if (response.ok) {
-          const lobbySocket: Socket = socketClass.getLobbySocket();
-          const player = Lobby.getPlayer();
-          lobbySocket.emit("update_gameobject", {
-            className: "Character",
-            objectId: player.objectId,
-            name: player.name,
-            x: player.x,
-            y: player.y,
-            avatar: user.avatar,
-            nickname: user.nickname,
-            animation: { name: player.animation.name, isStop: false },
-          });
+          return true;
         }
       })
-      .catch((err) => console.error(err));
+      .catch( function (err) {
+        console.error(err)
+        return false;
+      } );
+      if (response != undefined)
+        return response;
+      else
+        return false;
   }
 
   async function updateNickname(newNickname: string): Promise<boolean> {
